fix(rubocop): extract synchronizer update workflow
Move apply, rollback, and validation helpers out of TariffSynchronizer
into dedicated modules so Metrics method/module length can be approached
without changing CDS/TARIC apply or rollback behaviour.

Issue: No ticket/issue
